Refactor import paths for AssessmentType in CategoryItem, CompetencyItem, and EvaluationMapping components
