+++ clients/assessment_component/src/assessment/pages/SettingsPage/components/CategoryList/components/CompetencyItem/components/EvaluationMapping.tsx
@@ -12,7 +12,7 @@ import {
   TooltipProvider,
 } from '@tumaet/prompt-ui-components'
 
-import { AssessmentType } from '../../../../../interfaces/assessmentType'
+import { AssessmentType } from '../../../../../../../interfaces/assessmentType'
 import { Competency } from '../../../../../../../interfaces/competency'
 
 import { useUpdateCompetencyMapping } from '../hooks/useUpdateCompetencyMapping'
